Refactor of file locations
