+++ src/main/resources/com/example/main_sem_proj/view/hello-view.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.HBox?>
 
 <HBox alignment="CENTER" spacing="10" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.main_sem_proj.controller.authentication.HelloController">
+      fx:controller="com.example.main_sem_proj.controller.HelloController">
   <padding>
     <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
   </padding>
