+++ src/main/resources/com/example/main_sem_proj/view/login-view.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.control.Hyperlink?>
 <VBox xmlns:fx="http://javafx.com/javafx"
-      fx:controller="com.example.main_sem_proj.controller.authentication.LoginController"
+      fx:controller="com.example.main_sem_proj.controller.LoginController"
       alignment="CENTER"
       spacing="5">
