+++ src/main/resources/com/example/main_sem_proj/view/main-view.fxml
@@ -13,7 +13,7 @@
 
 <?import javafx.scene.layout.AnchorPane?>
 <VBox xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.main_sem_proj.controller.mainGUI.MainController"
+      fx:controller="com.example.main_sem_proj.controller.MainController"
       stylesheets="@../css/styles.css"
       alignment="CENTER">
