+++ src/test/java/MainControllerTest.java
@@ -1,18 +1,14 @@
-import com.example.main_sem_proj.controller.mainGUI.MainController;
-import org.junit.jupiter.api.*;
-import static org.junit.jupiter.api.Assertions.*;
-
 public class MainControllerTest {
 
-    private MainController controller;
-
-    @BeforeEach
-    void setUp() {
-        controller = new MainController(); // Initialize
-    }
-
-    @Test
-    void testSignout() {
-
-    }
+//    private MainController controller;
+//
+//    @BeforeEach
+//    void setUp() {
+//        controller = new MainController(); // Initialize
+//    }
+//
+//    @Test
+//    void testSignout() {
+//
+//    }
 }
