+++ src/test/java/NotificationsTest.java
@@ -1,4 +1,4 @@
-import com.example.main_sem_proj.controller.mainGUI.NotificationsController;
+import com.example.main_sem_proj.controller.NotificationsController;
 import com.example.main_sem_proj.model.Notifications.MockDisplayComponent;
 import org.junit.jupiter.api.*;
 import static org.junit.jupiter.api.Assertions.*;
